Fixed Open-RL Bug.

+++ CybORG/CybORG/Agents/Wrappers/OpenAIGymWrapper.py
@@ -28,13 +28,13 @@ class OpenAIGymWrapper(Env, BaseWrapper):
         result.observation = self.observation_change(result.observation)
         result.action_space = self.action_space_change(result.action_space)
         info = vars(result)
-        return np.array(result.observation), np.float32(result.reward), result.done, info
+        return np.array(result.observation, dtype=np.float32), result.reward, result.done, info
 
     def reset(self, agent=None):
         result = self.env.reset(self.agent_name)
         result.action_space = self.action_space_change(result.action_space)
         result.observation = self.observation_change(result.observation)
-        return np.array(result.observation)
+        return np.array(result.observation, dtype=np.float32)
 
     def render(self):
         # TODO: If FixedFlatWrapper it will error out!
